% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Ntm.R
\name{calc_Ntm}
\alias{calc_Ntm}
\title{Calculate Ntm parameter}
\usage{
calc_Ntm(obs, Ntm = seq(1, 100, 1))
}
\arguments{
\item{obs}{The observed commmunity matrix.}

\item{Ntm}{A starting sequence of integers used in finding Ntm.}
}
\value{
A list consiting of Ntm and the fit neutral matrix.
}
\description{
Finds the Ntm parameter by reiteration.
}
\details{
Calculates the expected probability of detection of each OTU in
 the target community if it was present merely because of neutral
  processes from the source community.
  These probabilities are calculated for all values of Ntm and then
  the best fit candidate is determined based upon a least sum of squares
  approach. The function is called recursively until Ntm is estimated
  to 2 decimal places.
}
\keyword{internal}
