% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_occur_abund.R
\name{freq_occur_abund}
\alias{freq_occur_abund}
\title{Calculate abundance statistics}
\usage{
freq_occur_abund(otu)
}
\arguments{
\item{otu}{A community matrix.}
}
\value{
A named list.
}
\description{
Calculate abundance statistics
}
\keyword{internal}
