% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model.R
\name{neutral_model}
\alias{neutral_model}
\alias{neutral_mode}
\title{Neutral Model}
\usage{
neutral_model(physeq, source.otu, target.otu, add_taxa = FALSE)
}
\arguments{
\item{physeq}{A phyloseq object containing samples for the source and target communities.}

\item{source.otu}{A matrix of counts with samples in rows and OTUs in columns.}

\item{target.otu}{A matrix of counts with samples in rows and OTUs in columns.}

\item{add_taxa}{A logical. If TRUE, taxa from the phyloseq object are included in the output.}
}
\value{
An S3 object of class NCM.
}
\description{
Fits OTU frequencies in the target community to OTU relative abundances in the source
  community. Returns an S3 object of class NCM with print, summary, and plot functions.
}
\details{
Returns an S3 object of class NCM with print, summary, and plot functions.
  The object consists of a list of 8 items:
  \enumerate{
  \item stats: Ntm, mean.Nt, amd m.
  \item obs: Statistics used by the plot function.
  \item bestneutralmatrix: tatistics used by the plot function.
  \item neutral_matrix: For each OTU, statistics on frequencies of occurrence and relative abundances
     for OTUs occurring in the target community as expected according to the neutral model. May also
     include taxonomic assignment of each OTU.
  \item against_matrix: Same data for OTUs occurring in the target community less frequently than
     expected.
  \item for_matrix: Same data for OTUs occurring in the target community more frequently than expected.
  \item sum1: Summary statistics included by the summary function.
  \item sum2: Summary statistics included by the summary function.
  }
}
\examples{
data(schmidt)
source.com <- veganotu(subset_samples(schmidt, Site=="Mouth"))
target.com <- veganotu(subset_samples(schmidt, Site=="Lung"))
test <- neutral_model(schmidt, source.com, target.com)
}
\author{
John Quensen, Tom Schmidt & Arvind Venkataraman
}
