% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NCM.R
\name{plot.NCM}
\alias{plot.NCM}
\alias{plot}
\title{Plot method for class NCM}
\usage{
plot.NCM(x, xlab = "Relative Abundance in Source",
  ylab = "Det.Freq. in Target", main = NULL,
  xlim = c(0.001, 100), ylim = c(0, 100), ...,
  symb.col = c("darkgrey", "indianred", "darkgreen"))
}
\arguments{
\item{x}{Output of the function neutral_model.}

\item{xlab}{Y-axis label.}

\item{ylab}{X-axis label.}

\item{main}{Plot title.}

\item{xlim}{X-axis limits.}

\item{ylim}{Y-axis limits}

\item{...}{Other parmeters passed to the generic plot function.}

\item{symb.col}{Vector of symbol colors for neutral, selected against, and selected for OTUs.}
}
\value{
A plot.
}
\description{
Plots an NCM object.
}
\examples{
data(rslt)
plot(rslt, xlab="Relative Abundance in Mouth", ylab="Det. Freq in Lung")

}
\author{
John Quensen
}
