% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NCM.R
\name{print.NCM}
\alias{print.NCM}
\title{Print method for class NCM}
\usage{
\method{print}{NCM}(x, ...)
}
\arguments{
\item{x}{Output of the function neutral_model.}

\item{...}{Other parameters passed to the generic print function.}
}
\value{
Printed summary.
}
\description{
Prints statistics for a class NCM object.
}
\examples{
data(rslt)
rslt

}
\author{
John Quensen
}
