% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NCMpval.R
\name{print.NCMpval}
\alias{print.NCMpval}
\title{Plot method for class NCMpval}
\usage{
\method{print}{NCMpval}(x, ...)
}
\arguments{
\item{x}{Output of the function pval_NeutralModel.}

\item{...}{Other parameters passed to the generic print function.}
}
\value{
Printed summary.
}
\description{
Prints a summary for a class NCMpval object.
}
\author{
John Quensen
}
