% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_NeutralModel.R
\name{pval_NeutralModel}
\alias{pval_NeutralModel}
\title{pval_NeutralModel}
\usage{
pval_NeutralModel(nm.rslt, n.iter=99999, plot=TRUE)
}
\arguments{
\item{nm.rslt}{Result from neutral_model function.}

\item{n.iter}{Number of iterations.}

\item{plot}{A logical; TRUE plots simulated distribution, with red line denoting observed statistic.}
}
\value{
A list of two items: the observed RMS error and associated p.values.
}
\description{
Makes a test of signifcance for the neutral model.
}
\details{
This test is not really valid. Non-linear models can only be tested against each other
  to see which is better. All are signficant by this method.
}
\examples{
data(rslt)
pval_NeutralModel(rslt)

}
\author{
Arvind Venkataraman and John Quensen
}
