% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rplb}
\alias{rplb}
\title{rplB gene count data from a pouchitis study}
\format{A phyloseq object containing OTU, sample data, and taxonomy tables.}
\usage{
rplb
}
\description{
An experiment level phyloseq object including rplB gene count data from a pouchitis study
}
\keyword{datasets}
