% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schmidt}
\alias{schmidt}
\title{schmidt data}
\format{An experiment level \code{phyloseq} object.}
\usage{
schmidt
}
\description{
Data from Schmidt paper as a phyloseq object containing OTU, sample data and taxonomy tables
}
\keyword{datasets}
