% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NCM.R
\name{summary.NCM}
\alias{summary.NCM}
\title{Summary method for class NCM}
\usage{
\method{summary}{NCM}(object, ...)
}
\arguments{
\item{object}{Output of the function neutral_model.}

\item{...}{Other parameters passed to the generic print function.}
}
\value{
A printed summary of the output of the function neutral_model.
}
\description{
Prints a summary for an NCM class object.
}
\examples{
data(rslt)
summary(rslt)
}
\author{
John Quensen
}
