% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veganotu.R
\name{veganotu}
\alias{veganotu}
\title{veganotu}
\usage{
veganotu(physeq)
}
\arguments{
\item{physeq}{A phyloseq object containing an OTU table.}
}
\value{
A community matrix with samples as rows and taxa as columns.
}
\description{
Extracts a community matrix from a phyloseq object.
}
\examples{
data(rplb)
veganotu(rplb)

}
\author{
John Quensen
}
